🟡 Player Conversion
- Player has been converted to an entity.
- Move some of the logic from player to entity base class.
- Store pointer to the derived animation class to update it.

+++ FlashFire/source/game/entities/entity.cpp
@@ -2,7 +2,7 @@
 
 namespace FF
 {
-    void entity::SetPosition(int newX, int newY)
+    void entity::SetPosition(float newX, float newY)
     {
         m_Position.x = newX;
         m_Position.y = newY;
